refactor: move common types to `lib/types/`
