+++ lib/types/req.ts
@@ -98,3 +98,21 @@ export const UpdateEpisodeComment = t.Object(
   },
   { $id: 'UpdateEpisodeComment' },
 );
+
+export const EpisodeExpected = t.Optional(
+  t.Partial(
+    t.Object(
+      {
+        name: t.String(),
+        nameCN: t.String(),
+        duration: t.String(),
+        date: t.String(),
+        summary: t.String(),
+      },
+      {
+        description:
+          "a optional object to check if input is changed by others\nif some key is given, and current data in database doesn't match input, subject will not be changed",
+      },
+    ),
+  ),
+);
